+++ lib/domain/state/app_state.dart
@@ -1,8 +1,0 @@
-import 'package:osam/domain/state/base_state.dart';
-
-// ignore: must_be_immutable
-abstract class AppState extends BaseState {
-  @override
-  // TODO: implement props
-  List<Object> get props => null;
-}