+++ lib/domain/store/store.dart
@@ -1,69 +1,71 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
-import 'package:osam/data/persist_repository.dart';
+import 'package:osam/persist/persist_interface.dart';
+import 'package:osam/persist/persist_repository.dart';
 import 'package:osam/domain/event/event.dart';
 import 'package:osam/domain/middleware/middleware.dart';
 import 'package:osam/domain/state/base_state.dart';
 
-abstract class Store<ST extends BaseState<ST>> {
+abstract class Store<ST extends BaseState<ST>> implements Persist {
+  factory Store(ST state, {List<Middleware<Store<BaseState<ST>>>> middleWares = const [], bool logging = false}) =>
+      _StoreImpl(appState: state, middleWares: middleWares, logging: logging);
+
   ST get state;
 
-  factory Store(ST state, {List<Middleware<Store<BaseState<ST>>>> middleWares = const []}) =>
-      _StoreImpl(appState: state, middleWares: middleWares);
+  bool get isPersistEnabled;
 
   Stream<ST> get nextState;
 
-  void dispatchEvent({@required Event<ST> event});
-
   Stream<Event<ST>> get eventStream;
 
-  Future<void> initPersist();
-
-  void storeState();
-
-  void wipePersist();
-
-  void restoreState();
+  void dispatchEvent<BT extends Object>({@required Event<ST> event});
 }
 
 class _StoreImpl<ST extends BaseState<ST>> implements Store<ST> {
   ST appState;
-
-  @override
-  ST get state => appState;
-
   final List<Middleware<Store<BaseState<ST>>>> middleWares;
+  final _denormalizedConditions = <Condition>[];
+  final bool logging;
 
   // ignore: close_sinks
   StreamController<Event<ST>> _dispatcher;
 
-  @override
-  Stream<Event<ST>> get eventStream => _dispatcher.stream;
-
-  final _denormalizedConditions = <Condition>[];
-
-  _StoreImpl({this.appState, this.middleWares}) {
+  _StoreImpl({this.appState, this.middleWares, this.logging}) {
     _initStore();
   }
+
+  @override
+  ST get state => appState;
+
+  @override
+  bool get isPersistEnabled => PersistRepository().isEnabled;
+
+  @override
+  Stream<ST> get nextState => appState.stateStream;
+
+  @override
+  Stream<Event<ST>> get eventStream => _dispatcher.stream;
 
   @override
   Future<void> initPersist() async => await PersistRepository().init();
 
   @override
-  void restoreState() => this.appState = PersistRepository().restoreState() ?? this.appState;
+  void dispatchEvent<BT extends Object>({@required Event<ST> event}) => _dispatcher.sink.add(event);
 
   @override
   void storeState() => PersistRepository().storeState(appState);
 
   @override
-  void wipePersist() => PersistRepository().wipeState();
+  void restoreState() => this.appState = PersistRepository().restoreState() ?? this.appState;
+
+  @override
+  void deleteState() => PersistRepository().deleteState();
 
   void _initStore() {
     _dispatcher = StreamController<Event<ST>>.broadcast();
     middleWares.forEach((middleWares) => middleWares.store = this);
-    _denormalizedConditions
-        .addAll(middleWares.expand((middleWare) => middleWare.conditions).toList());
+    _denormalizedConditions.addAll(middleWares.expand((middleWare) => middleWare.conditions).toList());
     _dispatcher.stream.listen((event) {
       for (Condition condition in _denormalizedConditions) {
         final nextEvent = condition(event);
@@ -72,7 +74,11 @@ class _StoreImpl<ST extends BaseState<ST>> implements Store<ST> {
         else
           break;
       }
+      if (logging)
+        debugPrint(
+            'Event in stores event stream is : ${'type: ' + event.type.toString() + ' bundle: ' + event.bundle.toString() + ' runtimeType: ' + event.runtimeType.toString()}');
       if (event is ModificationEvent) {
+        if (logging) debugPrint('reducer called : ${(event as ModificationEvent).reducer.toString}');
         try {
           event(appState, event.bundle);
         } catch (e) {
@@ -82,9 +88,5 @@ class _StoreImpl<ST extends BaseState<ST>> implements Store<ST> {
     });
   }
 
-  @override
-  Stream<ST> get nextState => appState.stateStream;
 
-  @override
-  void dispatchEvent({@required Event<ST> event}) => _dispatcher.sink.add(event);
 }